Finish migration to JDK7
Multicatch, for-loop, switch over strings, try catch, etc.

+++ seqware-portal/src/main/java/net/sourceforge/solexatools/webapp/controller/StudyReportTableController.java
@@ -529,7 +529,7 @@ public class StudyReportTableController extends BaseCommandController {
             List<String> cellsModel = new LinkedList<>();
             cellsModel.add(wrapSwAccession(rootSample.getSwAccession(), rootSample.getTitle()));
             cellsModel.add(wrapSwAccession(sample.getSwAccession(),
-                    rootSample.getSampleId().intValue() != sample.getSampleId().intValue() ? sample.getTitle() : "no child"));
+                    rootSample.getSampleId() != sample.getSampleId() ? sample.getTitle() : "no child"));
             cellsModel.addAll(statusesOut);
             cellsModel.add(wrapOverall());
